feat: Add new widgets and services for stationary dashboard
- Introduced EmptyStateWidget to display when no sensor readings are available.
- Created ReadingCard widget for displaying individual sensor readings.
- Added SensorConfigBar for sensor configuration options in the dashboard.
- Implemented ZoneCard to show readings for each zone.
- Developed AppStateService for managing application state and operations.
- Created DialogHelper for standardized dialog operations across the app.
- Introduced FormattersService for consistent data formatting utilities.
- Added ValidatorsService for centralized validation utilities.
