+++ application/agrivoltaics_flutter_app/lib/pages/stationary_dashboard/widgets/config_row.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/material.dart';
+
+// Small row for displaying key/value pairs in dialogs
+class ConfigRow extends StatelessWidget {
+  final String label;
+  final String value;
+
+  const ConfigRow({required this.label, required this.value});
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.only(bottom: 8),
+      child: Row(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          SizedBox(
+            width: 130,
+            child: Text(
+              label,
+              style: const TextStyle(fontWeight: FontWeight.w600),
+            ),
+          ),
+          Expanded(
+            child: SelectableText(value),
+          ),
+        ],
+      ),
+    );
+  }
+}
